requirements.txt
django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
gunicorn==21.2.0
python-decouple==3.5
django-filter==21.1